requirements.txt
absl-py==2.1.0
altair==5.5.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==23.2.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.1
certifi==2021.5.30
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.4
contourpy==1.2.1
cvzone==1.6.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.2.1
django-cors-headers==4.7.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.53.0
fqdn==1.5.1
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
gTTS==2.5.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==2.10
ifaddr==0.2.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keyboard==0.13.5
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.1.1
ml-dtypes==0.4.0
mlxtend==0.23.4
msvc_runtime==14.40.33807
mysql-connector-python==9.2.0
narwhals==1.27.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.0.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==10.3.0
platformdirs==4.3.6
playsound==1.3.0
plotly==6.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==4.25.3
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
PyAudio==0.2.14
pycparser==2.22
pydeck==0.9.1
pyFirmata==1.1.0
pygame==2.6.0
Pygments==2.19.1
pyogrio==0.10.0
pyparsing==3.1.2
pypiwin32==223
pyproj==3.7.0
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.10.0
pyserial==3.5
pysmb==1.2.11
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pyttsx3==2.90
pytz==2025.1
pywin32==306
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
qrcode==8.2
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
shapely==2.0.7
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.4.7
soupsieve==2.6
SpeechRecognition==3.10.4
speedtest==0.0.1
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.42.1
tenacity==9.0.0
terminado==0.18.1
tesseract==0.1.3
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==1.26.6
virtualenv==20.31.1
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
zeroconf==0.147.0
